clase2/Practica/python-graphql-docker/pyproject.toml
[project]
name = "python-graphql-docker"
version = "0.1.0"
description = ""
authors = [
    {name = "gasper-az",email = "acevedo.zain.gaspar@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "strawberry-graphql (>=0.311.1,<0.312.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
